Fixed #80 Bug
修复 #80 Bug
